requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.0.3
python-dateutil==2.8.2
python-barcode==0.15.1
Pillow==10.1.0
requests==2.31.0
Flask-Login==0.6.3
APScheduler==3.10.4
pywebpush==1.14.0
